% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age.R
\name{add_age}
\alias{add_age}
\title{Derive age and age band from year of birth}
\usage{
add_age(data, write = FALSE, measurefrom = Sys.time())
}
\arguments{
\item{data}{a data table of patient data including at least patient id and age}

\item{write}{boolean to save as parquet file}

\item{measurefrom}{date to measure patient age at (eg registration date)}
}
\value{
data.table() of patient data with age and age band appended
}
\description{
Derive age and age band from year of birth
}
