% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurum_pipeline.R
\name{aurum_pipeline}
\alias{aurum_pipeline}
\title{Aurum Pipeline}
\usage{
aurum_pipeline(
  type,
  cols = NULL,
  saveloc = "",
  dataloc,
  patids = NULL,
  check = FALSE
)
}
\arguments{
\item{type}{character string of Aurum table}

\item{cols}{character string of column types (see tabledata with meta data for this)}

\item{saveloc}{character string of location to save parquet files - defaults to project/working directory}

\item{dataloc}{character string of location of raw data}

\item{patids}{vector of patient ids to filter to (if known)}

\item{check}{boolean run in check mode or regular}
}
\value{
data.frame of data checking results
}
\description{
Aurum Pipeline
}
\examples{
\dontrun{
aurum_pipeline(type = 'Patient', 
dataloc = 'C:/Pathto/Your/CPRD/Extract/Here', check = TRUE)}
}
