% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_patid_links.R
\name{check_patid_links}
\alias{check_patid_links}
\title{Check patients id links}
\usage{
check_patid_links(
  dataloc = "",
  dts = c("Consultation", "Observation", "Referral", "DrugIssue", "Problem")
)
}
\arguments{
\item{dataloc}{character filepath to parquet data folder (default to project directory)}

\item{dts}{character vector of activities to check}
}
\value{
Results of link tests as data.frame()
}
\description{
Check patients id links
}
\examples{
\dontrun{
check_patid_links()}
}
