% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars.R
\name{check_vars}
\alias{check_vars}
\title{Check Aurum Variables}
\usage{
check_vars(table, data, chart = FALSE)
}
\arguments{
\item{table}{character vector of tables to check (subset of tabledata$table_name)}

\item{data}{output of Aurum_Pipeline()}

\item{chart}{boolean to create charts of pipeline output}
}
\value{
list of summarised table checks
}
\description{
Check Aurum Variables
}
\examples{
\dontrun{
check_vars(c('Patient', 'Consultation', 'Observation'), res)}
}
