% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_medcodes.R
\name{cond_medcodes}
\alias{cond_medcodes}
\title{Conditions from medcodes}
\usage{
cond_medcodes(diag_obs, codelist, var_name, enddate)
}
\arguments{
\item{diag_obs}{observation data}

\item{codelist}{codelist of conditions}

\item{var_name}{column to read conditions from in codelist}

\item{enddate}{date to record conditions to}
}
\value{
data.table() of medcodes and conditions
}
\description{
Conditions from medcodes
}
\examples{
\dontrun{
cond_medcodes(diag_obs = diag_obs
, codelist = codelist, var_name = 'Colname', enddate = '2016-01-01')}
}
