% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_codes}
\alias{get_codes}
\title{Get (med/prod)codes from data}
\usage{
get_codes(
  dataloc = "",
  startdate = "1800-01-01",
  enddate,
  recs = NULL,
  obsflag = T,
  codelist
)
}
\arguments{
\item{dataloc}{path to aurum data files}

\item{startdate}{date to include observations / drug issues from}

\item{enddate}{date to include observations / drug issues to}

\item{recs}{observation or drug issue data if already in environment}

\item{obsflag}{flag to indicate observation (T) or drug issue (F)}

\item{codelist}{existing codelist (medcodes for obsevations, prodcodes for drug issues)}
}
\value{
data.table() of relevent codes
}
\description{
Get (med/prod)codes from data
}
\examples{
\dontrun{
diag_obs_test <- get_codes(enddate = '2016-01-01', codelist = medcodelist)}
}
