% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenDT.R
\name{opendt}
\alias{opendt}
\title{Open Data Table}
\usage{
opendt(
  data_in,
  cols_in = NULL,
  date_in = NULL,
  start_date = NULL,
  end_date = NULL,
  patient_list = NULL
)
}
\arguments{
\item{data_in}{character filepath to data or an S3 URI. If data saved in standard structure from aurum_pipeline()
then only table name is required, eg 'Observation', otherwise use full file path.}

\item{cols_in}{character vector of column names to load in}

\item{date_in}{character name of date field to filter on}

\item{start_date}{Date to filter from (inclusive)}

\item{end_date}{Date to filter to (inclusive)}

\item{patient_list}{Vector of patient ids to filter to}
}
\value{
A data.table()
}
\description{
Open Data Table
}
\examples{
\dontrun{
opendt('Patient', date_in = 'regstartdate', start_date = '2020-01-01', end_date = '2020-01-31')}
}
