% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{path_id}
\alias{path_id}
\title{Filepath Identifier}
\usage{
path_id(f_path)
}
\arguments{
\item{f_path}{character}
}
\value{
vector
}
\description{
Filepath Identifier
}
\examples{
path_id('s3://thf-dap-tier3-projects-test123-8681-projectbucket-di2wwm3vl5fz/Data/Patient')
}
