% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multimorb_codelists.R
\name{read_multimorb_codelists}
\alias{read_multimorb_codelists}
\title{Read in Multimorbidity Codelists}
\usage{
read_multimorb_codelists(codeloc, file_pattern = NULL, coltypes = NULL)
}
\arguments{
\item{codeloc}{Either a character string filepath to folder containing the codelists, or an s3 URI}

\item{file_pattern}{Use only the codelists with this in their name (currently non-s3 only)}

\item{coltypes}{If data types are known}
}
\value{
data.table() of combined codelists
}
\description{
Read in Multimorbidity Codelists
}
