% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabledata}
\alias{tabledata}
\title{Metadata about CPRD table structure}
\format{
A tibble with 8 rows and 3 variables:
\describe{
\item{table_name}{The tables from a CPRD Aurum extract}
\item{cols}{The data types of the variables (from the DFPC extract)}
\item{main_date}{The most commonly used date for each table (where it exists)}
}
}
\usage{
tabledata
}
\description{
A dataset with information on CPRD tables and variable formats
}
\keyword{datasets}
